requirements.txt
    Pillow
    Django
    djangorestframework
